refactor(promptbox): decompose into three focused hooks
PromptBox.tsx was the heaviest, highest-traffic file in the codebase
(~500 lines after image paste) — every recent feature ships through
it (@-mentions, paperclip, paste, thumbnails, lightbox, chip
backdrop, IME guards). Pull three independent concerns into focused
hooks so the component shrinks and each piece is testable in
isolation:

  - `usePromptText` (49 lines, webview/src/hooks/usePromptText.ts) —
    Owns text state, textarea + backdrop refs, auto-resize on text
    change, and the post-mutation caret-restore pattern via
    `pendingCursor`. The `TEXTAREA_MAX_HEIGHT = 160` constant lives
    here now (kept in sync with the `--message-max-height` CSS var
    via comments on both sides).
  - `useImageAttachments` (50 lines,
    webview/src/hooks/useImageAttachments.ts) — Owns the image
    thumbnail strip + lightbox preview state. Initial seeding from
    `initial.attachments` (image-mime only) happens in a useState
    lazy initializer. Exposes `addImages`,
    `removeImageAttachment`, `clearImageAttachments`,
    `previewImage`, `setPreviewImage`.
  - `useMentionPicker` (88 lines,
    webview/src/hooks/useMentionPicker.ts) — Owns mention / hits /
    activeIndex state, the search-debouncing effect, and the
    `insertMention` callback (which also mutates the parent's
    `knownMentions` ref and stages `pendingCursor`, passed in as
    refs to avoid copy-then-sync).

PromptBox itself is now ~91 lines shorter and reads as: text state +
chip-attachment state + submit + JSX. Each hook is independent and
can grow without bloating the component.

No user-visible change. 484 tests still pass.
